Significant readability improvement
Somewhere in my initial testing, I ran into an error when using the += syntax, or maybe it was the increment (++) syntax. So I stopped using it. But then I looked into it further and found that it works just fine and the error must have been for some other reason. So now I'm using that instead, because it's so much easier to read.